ePiphany/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
frozenlist==1.4.0
halo==0.0.31
idna==3.4
log-symbols==0.0.14
multidict==6.0.4
openai==0.27.8
phpserialize==1.3
psycopg2-binary==2.9.7
requests==2.31.0
simplejson==3.19.1
six==1.16.0
spinners==0.0.24
termcolor==2.3.0
tqdm==4.66.1
urllib3==2.0.4
WooCommerce==3.0.0
yarl==1.9.2
